Create setupApp
Move the global ValidationPipe and sessions setup to setup-app.ts
